moved/changed some files and namespaces around, fixed logger config not loading
