+++ onboard/godot-frontend/app.config
@@ -2,7 +2,7 @@
     <configSections>
         <section name="log4net" type="log4net.Config.Log4NetConfigurationSectionHandler, log4net" />
     </configSections>
-    <log4net>
+    <log4net debug="false">
         <appender name="FileAppender" type="log4net.Appender.FileAppender">
             <file type="Log4net.Util.PatternString" value="%property{LogFilePath}/%date{yyyy'-'MM'-'dd_HH':'mm':'ss}%property{LogFileName}"/>
             <appendToFile value="true" />
