+++ onboard/godot-frontend/main.tscn
@@ -1,6 +1,6 @@
 [gd_scene load_steps=129 format=3 uid="uid://br86segmlm7yg"]
 
-[ext_resource type="Script" uid="uid://cpnqnrrdr0pbv" path="res://GuiManager.cs" id="1_ig7tw"]
+[ext_resource type="Script" uid="uid://cpnqnrrdr0pbv" path="res://GUIs/GuiManager.cs" id="1_ig7tw"]
 [ext_resource type="PackedScene" uid="uid://bs2l3u3gpwk3a" path="res://GUIs/orignial/Original.tscn" id="2_lquwl"]
 [ext_resource type="Texture2D" uid="uid://cm8hlcb1in1rd" path="res://CSHAssets/transparent-logo-white.png" id="3_1bvp3"]
 [ext_resource type="SpriteFrames" uid="uid://cwsn76nuho5hi" path="res://CSHAssets/loadingAnimation.tres" id="3_h2yge"]
