+++ onboard/godot-frontend/project.godot
@@ -16,6 +16,10 @@ config/features=PackedStringArray("4.4", "C#", "GL Compatibility")
 run/low_processor_mode=true
 config/icon="res://icon.svg"
 
+[autoload]
+
+LogConfigAutoLoad="*res://LogConfigAutoLoad.cs"
+
 [display]
 
 window/size/viewport_width=900
